+++ front/src/componentes/Footer.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import './Footer.css'
+import '../styles/Footer.css'
 
 export default function Footer() {
   return (
