+++ front/src/componentes/TopBar.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Form, Link, Navigate } from "react-router-dom"
-import './TopBar.css'
+import '../styles/TopBar.css'
 
 export default function TopBar() {
   return (
