+++ front/src/styles/App.css
